chore: move PrimerKlarnaPaymentView code to SDK
