--- packages/example/src/components/PrimerKlarnaPaymentView.tsx
@@ -1,6 +1,0 @@
-import { requireNativeComponent } from 'react-native';
-
-// TODO TWS-94: define in SDK?
-export const PrimerKlarnaPaymentView = requireNativeComponent<{}>(
-    'PrimerKlarnaPaymentView'
-);
