+++ packages/sdk/src/HeadlessUniversalCheckout/Components/PrimerKlarnaPaymentViewRaw.tsx
@@ -1,0 +1,5 @@
+import { requireNativeComponent } from "react-native";
+
+export const PrimerKlarnaPaymentViewRaw = requireNativeComponent<{}>(
+    'PrimerKlarnaPaymentView'
+);
